refactor(qa): centralize QA target encoding
