+++ src/plugin-sdk/qa-channel-protocol.ts
@@ -11,6 +11,18 @@ export type QaTargetParts = {
   threadId?: string;
 };
 
+/** Encode a canonical QA channel target. */
+export function buildQaTarget(params: {
+  chatType: QaBusConversationKind;
+  conversationId: string;
+  threadId?: string | null;
+}): string {
+  if (params.threadId) {
+    return `thread:${params.conversationId}/${params.threadId}`;
+  }
+  return `${params.chatType === "direct" ? "dm" : params.chatType}:${params.conversationId}`;
+}
+
 /** Parse the lowercase, prefix-scoped target grammar shared by QA Channel and QA Lab. */
 export function parseQaTarget(
   raw: string,
